+++ eslint.config.js
@@ -13,6 +13,7 @@ export default [
       globals: {
         ...globals.browser,
         ...globals.node,
+        NodeJS: false,
       },
     },
   },